django2.2/Advancing-the-Blog/requirements.txt
Django==2.2
django-crispy-forms==1.9.2
django-filter==2.3.0
django-markdown-deux==1.0.5
django-pagedown==2.1.2
djangorestframework==3.11.1
importlib-metadata==1.7.0
Markdown==3.2.2
markdown2==2.3.9
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
